+++ crates/zizmor/tests/integration/e2e.rs
@@ -408,21 +408,21 @@ fn issue_1065() -> Result<()> {
        |
     12 | /   issue-1065:
     13 | |     runs-on: ubuntu-latest
-    14 | |     steps:
-    15 | |       - name: Comment PR
+    14 | |     environment: foobar
+    15 | |     steps:
     ...  |
-    24 | |             Please review the changes and provide any feedback. Thanks! 🚀
-       | |                                                                          ^
-       | |                                                                          |
-       | |__________________________________________________________________________this job
-       |                                                                            default permissions used due to no permissions: block
+    25 | |             Please review the changes and provide any feedback. Thanks! 🚀
+       | |                                                                           ^
+       | |                                                                           |
+       | |___________________________________________________________________________this job
+       |                                                                             default permissions used due to no permissions: block
        |
        = note: audit confidence → Medium
 
     error[unpinned-uses]: unpinned action reference
-      --> @@INPUT@@:16:15
+      --> @@INPUT@@:17:15
        |
-    16 |         uses: thollander/actions-comment-pull-request@v3
+    17 |         uses: thollander/actions-comment-pull-request@v3
        |               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ action is not pinned to a hash (required by blanket policy)
        |
        = note: audit confidence → High
@@ -660,7 +660,6 @@ fn test_github_output() -> Result<()> {
     ::error file=@@INPUT@@,line=5,title=excessive-permissions::several-vulnerabilities.yml:5: overly broad permissions: uses write-all permissions
     ::error file=@@INPUT@@,line=11,title=excessive-permissions::several-vulnerabilities.yml:11: overly broad permissions: uses write-all permissions
     ::error file=@@INPUT@@,line=2,title=dangerous-triggers::several-vulnerabilities.yml:2: use of fundamentally insecure workflow trigger: pull_request_target is almost always used insecurely
-    ::warning file=@@INPUT@@,line=16,title=template-injection::several-vulnerabilities.yml:16: code injection via template expansion: may expand into attacker-controllable code
     ::error file=@@INPUT@@,line=16,title=template-injection::several-vulnerabilities.yml:16: code injection via template expansion: may expand into attacker-controllable code
     ::warning file=@@INPUT@@,line=1,title=concurrency-limits::several-vulnerabilities.yml:1: insufficient job-level concurrency limits: missing concurrency setting
     "